+++ diamondback/transforms/__init__.py
@@ -13,9 +13,9 @@
 
 from .ComplexTransform import ComplexTransform
 from .FourierTransform import FourierTransform
-from .PowerSpectrumTransform import PowerSpectrumTransform
+from .PsdTransform import PsdTransform
 from .WaveletTransform import WaveletTransform
 from .ZTransform import ZTransform
 
-__all__ = [ 'ComplexTransform', 'FourierTransform', 'PowerSpectrumTransform', 'WaveletTransform',
+__all__ = [ 'ComplexTransform', 'FourierTransform', 'PsdTransform', 'WaveletTransform',
             'ZTransform' ]